Rename file outline button
